move kubernetes deploy helpers to shared client mod
